fix debug spam fmt

+++ libclamav/pe.c
@@ -2560,7 +2560,7 @@ int cli_peheader(fmap_t *map, struct cli_exe_info *peinfo)
 				    if(v) {
 					s = cli_str2hex(vptr + 6, s_key_sz + s_val_sz - 6);
 					if(s) {
-					    cli_dbgmsg("VersionInfo (%x): '%s'='%s' - VI:%s\n", vptr - baseptr + 6, k, v, s);
+					    cli_dbgmsg("VersionInfo (%x): '%s'='%s' - VI:%s\n", (uint32_t)(vptr - baseptr + 6), k, v, s);
 					    free(s);
 					}
 					free(v);
